pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cryptotrader"
version = "0.1.0"
description = "Async multi-bot paper trading engine with MySQL and exchange abstraction."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Cryptotrader contributors" }]
dependencies = [
    "sqlalchemy[asyncio]>=2.0.36",
    "aiomysql>=0.2.0",
    "cryptography>=42.0.0",
    "greenlet>=3.0.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.12",
    "pre-commit>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["Database", "Engine", "Exchange", "Utils"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = [".", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]

[tool.ruff.lint.per-file-ignores]
# Sync pathlib for entrypoint is intentional in async `main`.
"main.py" = ["ASYNC240"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
show_error_codes = true
files = ["main.py", "Database", "Engine", "Exchange", "Utils", "tests"]
mypy_path = "."
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["aiomysql.*", "yaml.*", "tenacity.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.coverage.run]
source = ["Database", "Engine", "Exchange", "Utils"]
